+++ tdrs-frontend/src/actions/requestAccess.js
@@ -1,4 +1,5 @@
 import axios from 'axios'
+import { SET_AUTH } from './auth'
 
 export const PATCH_REQUEST_ACCESS = 'PATCH_REQUEST_ACCESS'
 export const SET_REQUEST_ACCESS = 'SET_REQUEST_ACCESS'
@@ -17,8 +18,9 @@ export const requestAccess = ({ firstName, lastName, stt: { id } }) => async (
     })
 
     if (data) {
+      dispatch({ type: SET_REQUEST_ACCESS })
       dispatch({
-        type: SET_REQUEST_ACCESS,
+        type: SET_AUTH,
         payload: { user: data },
       })
     } else {
