+++ tdrs-frontend/src/components/Request/index.js
@@ -1,0 +1,3 @@
+import Request from './Request'
+
+export default Request
