--- tdrs-frontend/src/components/Unassigned/index.js
@@ -1,3 +1,0 @@
-import Unassigned from './Unassigned'
-
-export default Unassigned
